Refactor CGQuery and use command classes
